refactor Modal file
